pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "percentile-grids-api"
version = "0.0.1"
description = "FastAPI-based GAMLSS percentile calculation service for neuroimaging data"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Mateusz Rakowski" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    "fastapi[standard]",
    "uvicorn[standard]",
    "python-multipart",
    "aiofiles",
    "pydantic",
    "pydantic-settings",
    "httpx",
    "pandas",
    "numpy",
    "matplotlib",
    "rpy2",
    "sse-starlette",
    "sqlmodel",
    "pyjwt[crypto]",
    "pwdlib[argon2]",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "httpx",
    "ruff",
    "mypy",
]

[project.scripts]
percentile-grids = "app.fastapi.main:app"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
